+++ tests/unit/distributed/dist_server_test.py
@@ -8,8 +8,8 @@ from gigl.distributed.graph_store import dist_server
 from gigl.distributed.graph_store.messages import (
     FetchABLPInputRequest,
     FetchNodesRequest,
-    InitSamplingBackendOpts,
-    RegisterBackendOpts,
+    InitSamplingBackendRequest,
+    RegisterBackendRequest,
 )
 from gigl.src.common.types.graph_data import Relation
 from tests.test_assets.distributed.test_dataset import (
@@ -580,7 +580,7 @@ class TestDistServerSampling(TestCase):
         runtime = mock_backend_cls.return_value
 
         backend_id_1 = self.server.init_sampling_backend(
-            InitSamplingBackendOpts(
+            InitSamplingBackendRequest(
                 backend_key="neighbor_loader_0",
                 worker_options=self.worker_options,
                 sampler_options=self.sampler_options,
@@ -588,7 +588,7 @@ class TestDistServerSampling(TestCase):
             )
         )
         backend_id_2 = self.server.init_sampling_backend(
-            InitSamplingBackendOpts(
+            InitSamplingBackendRequest(
                 backend_key="neighbor_loader_0",
                 worker_options=self.worker_options,
                 sampler_options=self.sampler_options,
@@ -609,7 +609,7 @@ class TestDistServerSampling(TestCase):
     ) -> None:
         runtime = mock_backend_cls.return_value
         backend_id = self.server.init_sampling_backend(
-            InitSamplingBackendOpts(
+            InitSamplingBackendRequest(
                 backend_key="neighbor_loader_0",
                 worker_options=self.worker_options,
                 sampler_options=self.sampler_options,
@@ -618,7 +618,7 @@ class TestDistServerSampling(TestCase):
         )
 
         channel_id = self.server.register_sampling_input(
-            RegisterBackendOpts(
+            RegisterBackendRequest(
                 backend_id=backend_id,
                 worker_key="neighbor_loader_0_compute_rank_0",
                 sampler_input=MagicMock(),
@@ -641,7 +641,7 @@ class TestDistServerSampling(TestCase):
     ) -> None:
         runtime = mock_backend_cls.return_value
         backend_id = self.server.init_sampling_backend(
-            InitSamplingBackendOpts(
+            InitSamplingBackendRequest(
                 backend_key="neighbor_loader_0",
                 worker_options=self.worker_options,
                 sampler_options=self.sampler_options,
@@ -649,7 +649,7 @@ class TestDistServerSampling(TestCase):
             )
         )
         channel_id = self.server.register_sampling_input(
-            RegisterBackendOpts(
+            RegisterBackendRequest(
                 backend_id=backend_id,
                 worker_key="neighbor_loader_0_compute_rank_0",
                 sampler_input=MagicMock(),
@@ -678,7 +678,7 @@ class TestDistServerSampling(TestCase):
     ) -> None:
         runtime = mock_backend_cls.return_value
         backend_id = self.server.init_sampling_backend(
-            InitSamplingBackendOpts(
+            InitSamplingBackendRequest(
                 backend_key="neighbor_loader_0",
                 worker_options=self.worker_options,
                 sampler_options=self.sampler_options,
@@ -686,7 +686,7 @@ class TestDistServerSampling(TestCase):
             )
         )
         channel_id = self.server.register_sampling_input(
-            RegisterBackendOpts(
+            RegisterBackendRequest(
                 backend_id=backend_id,
                 worker_key="neighbor_loader_0_compute_rank_0",
                 sampler_input=MagicMock(),
